migrate.py to run extension "fourcat_install.py" files
