+++ extensions/app_stores/requirements.txt
@@ -1,3 +1,2 @@
 itunes-app-scraper-dmi>=0.9.5
 google-play-scraper-dmi>=0.9.17
-selenium==4.3.0
